refactor: move getProductVersion into PackageWindowsHelper
